correct heating and cooling time

+++ slm_model/batch_solution.py
@@ -650,7 +650,7 @@ def calculate_batch_energy(
         [machine.power_subsystems["recoater_motor"]], 
         [machine.power_subsystems["electric_valves"]], 
         [machine.power_subsystems["gas_circulation_pump_motor"]], 
-    ], dtype=np.float64)
+    ], dtype=np.float128)
     
     # ==========================================================================
     # Building time array
@@ -663,12 +663,12 @@ def calculate_batch_energy(
         [batch_time_dict["support_building_time"]], 
         [batch_time_dict["recoater_time_all"]], 
         [batch_time_dict["cooling_time"]]
-        ], dtype=np.float64)
+        ], dtype=np.float128)
     
     # ==========================================================================
     # Calculating total energy and anergy matrix
     
-    K = machine.power_coefficient.values
+    K = machine.power_coefficient.values.astype(np.float128)
     # Equivalent to EPC = np.dot(np.dot(P, K), T.T)
     EPC = (P.T @ K @ T).reshape(-1).item() # An number
     
@@ -680,7 +680,7 @@ def calculate_batch_energy(
     #     for j in range(7):
     #         K_row.append(int(P[i] * K[i][j] * T[j]))
     #     PKT.append(K_row)
-    PKT = (P * K * T.T).astype(np.int64)
+    PKT = (P * K * T.T).astype(np.float64)
     
     energy_matrix = pd.DataFrame(
         data=PKT,
